Move IAO Dashboard to own folder
