+++ apps/forms-flow-ai/forms-flow-web/src/components/FOI/Dashboard/IAO/index.js
@@ -1,4 +1,3 @@
-
 import Dashboard from "./IAO/Dashboard";
 import MinistryDashboard from "./MinistryDashboard";
-export { Dashboard, MinistryDashboard };
+export { Dashboard };
